refactor: abstract API proxy to share defaults
