+++ src/util/server/apiProxy.js
@@ -1,0 +1,16 @@
+import httpProxyMiddleware from 'next-http-proxy-middleware'
+
+
+
+
+
+export default function apiProxy (...options) {
+  return (res, req) => {
+    return httpProxyMiddleware(req, res, {
+      changeOrigin: true,
+      secure: true,
+      xfwd: true,
+      ...options,
+    })
+  }
+}
